+++ src/main/java/roomescape/controller/RenderController.java
@@ -1,0 +1,18 @@
+package roomescape.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class RenderController {
+
+	@GetMapping("/")
+	public String renderHomePage() {
+		return "home";
+	}
+
+	@GetMapping("/reservation")
+	public String renderReservationPage() {
+		return "reservation";
+	}
+}
