+++ src/main/java/roomescape/controller/ReservationController.java
@@ -3,35 +3,25 @@ package roomescape.controller;
 import java.util.List;
 
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.RestController;
 
 import roomescape.dto.request.ReservationSaveRequest;
 import roomescape.dto.response.ReservationResponse;
 import roomescape.service.ReservationService;
 
-@Controller
-public class RoomEscapeController {
+@RestController
+public class ReservationController {
 
 	private final ReservationService reservationService;
 
-	public RoomEscapeController(ReservationService reservationService) {
+	public ReservationController(ReservationService reservationService) {
 		this.reservationService = reservationService;
-	}
-
-	@GetMapping("/")
-	public String renderHomePage() {
-		return "home";
-	}
-
-	@GetMapping("/reservation")
-	public String renderReservationPage() {
-		return "reservation";
 	}
 
 	@GetMapping("/reservations")
